Fixes location of high card size bits in V2.00 CSD
The SD card size is encoded in 22 bits in the V2.00 CSD block. Due to the way bits are numbered in the spec and C bitfields are laid out on little-endian systems, the 6-bit `c_size_high` field must be listed before the `reserved3` dummy field to include the low 6 bits of the byte.